+++ plugin/controllers/models/info.py
@@ -22,7 +22,7 @@ import os
 import sys
 import time
 
-OPENWEBIFVER = "OWIF 0.2.2"
+OPENWEBIFVER = "OWIF 0.2.3"
 
 def getOpenWebifVer():
 	return OPENWEBIFVER