+++ plugin/controllers/root.py
@@ -21,6 +21,7 @@ from AT import ATController
 from ER import ERController
 from BQE import BQEController
 from transcoding import TranscodingController
+from wol import WOLSetupController, WOLClientController
 from twisted.web import static, http
 
 class RootController(BaseController):
@@ -44,6 +45,8 @@ class RootController(BaseController):
 		self.putChild("epgrefresh", ERController(session))
 		self.putChild("bouqueteditor", BQEController(session))
 		self.putChild("transcoding", TranscodingController(session))
+		self.putChild("wol", WOLClientController(session))
+		self.putChild("wolsetup", WOLSetupController(session))
 		if piconpath:
 			self.putChild("picon", static.File(piconpath))